Change SCSS to LESS, add less-watch-compiler, separate less files for some components, re-work .addThemes()
