+++ homework_13_create-react-app/fish-shop/less-watcher.config.json
@@ -1,0 +1,7 @@
+{
+  "watchFolder": "src",
+  "outputFolder": "src",
+  "sourceMap": true,
+  "runOnce": false,
+  "enableJs": true
+}
