+++ homework_13_create-react-app/fish-shop/package.json
@@ -16,7 +16,7 @@
     "typescript": "^4.4.4"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "concurrently --kill-others \"less-watch-compiler --config less-watcher.config.json\" \"react-scripts start\"",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -40,6 +40,9 @@
     ]
   },
   "devDependencies": {
+    "concurrently": "^6.3.0",
+    "less": "^4.1.2",
+    "less-watch-compiler": "^1.16.3",
     "node-sass": "^6.0.1"
   }
 }
