+++ homework_13_create-react-app/fish-shop/src/App.less
@@ -1,0 +1,100 @@
+@import (reference) 'references/constants';
+@import (reference) 'references/silent-classes';
+@import (reference) 'references/mixin-functions';
+
+.page-layout:extend(.x-grid) {
+  height: 100vh;
+  grid-template-columns: 240px 1fr auto;
+  grid-template-rows: 57px auto 48px;
+  grid-template-areas:
+        ". . ."
+        "aside main trailing"
+        ". . .";
+  .addThemes(@backgroundColorLight, @backgroundColorDark; @contentColorLight, @contentColorDark);
+  &__header:extend(.x-flex) {
+    grid-column-end: span 3;
+    .addThemes(@headerColorLight @headerColorDark, @nameColorLight @nameColorDark);
+    &:hover~.page-layout__main {
+      .product-tile:not(:hover) {
+        transition-duration: 0s;
+      }
+    }
+  }
+  &__navigation {
+    align-self: flex-end;
+    display: flex;
+  }
+  &__nav-items {
+    list-style-type: none;
+    padding: 0;
+    margin: 0 0 3px 0;
+    width: 300px;
+    display: flex;
+    flex-wrap: nowrap;
+    justify-content: flex-end;
+  }
+  &__nav-item {
+    padding-right: @spacer - 2px;
+  }
+  &__filter-switch {
+    visibility: hidden;
+  }
+  &__main:extend(.x-flex-column) {
+    grid-area: main;
+    flex: none;
+  }
+  &__aside:extend(.x-flex-column) {
+    grid-area: aside;
+    &_closed {
+      display: none;
+    }
+  }
+  &__trailing {
+    grid-area: trailing;
+  }
+  &__footer:extend(.x-flex) {
+    grid-column-end: span 3;
+    .addThemes(@paneColorLight @paneColorDark);
+  }
+  @media @maxAux {
+    &__nav-items {
+      width: 140px;
+      flex-direction: column;
+      font-size: 14px;
+    }
+    &__trailing {
+      min-width: @halfSpacer;
+    }
+  }
+  @media @maxPhone {
+    grid-template-columns: 1fr 1fr auto;
+    grid-template-rows: 57px auto auto 48px;
+    grid-template-areas:
+      ". . ."
+      "main main trailing"
+      "aside aside trailing"
+      ". . .";
+    &__main,
+    &__aside {
+      margin-left: @halfSpacer;
+    }
+    &__filter-switch {
+      visibility: visible;
+    }
+    &__trailing {
+      min-width: 0;
+    }
+  }
+  @media @minmaxLsPad {
+    &__trailing {
+      min-width: 5vw;
+      max-width: 5vw;
+    }
+  }
+  @media @minDesktop  {
+    &__main {
+      min-width: 712px;
+      max-width: 712px;
+    }
+  }
+}
