+++ homework_13_create-react-app/fish-shop/src/components/product-filter/ProductFilter.less
@@ -1,0 +1,25 @@
+@import (reference) '../../references/constants';
+@import (reference) '../../references/silent-classes';
+@import (reference) '../../references/mixin-functions';
+
+.product-filter {
+  width: 100%;
+  height: 100%;
+  padding: @spacer 0;
+  .addThemes(@paneColorLight @paneColorDark);
+  &__type-block {
+    border: 0;
+  }
+  &__block-title {
+    padding: 0;
+    font-size: 16px;
+    font-weight: bold;
+  }
+  @media @maxPhone {
+    display: flex;
+    justify-content: space-between;
+    &__type-block {
+      padding: 0 @doubleSpacer;
+    }
+  }
+}
