+++ homework_13_create-react-app/fish-shop/src/components/product-tile/ProductTile.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import './ProductTile.css';
 import {ProductTileButtons} from "./__buttons/ProductTileButtons";
 
 export class ProductTile extends React.Component<any, any> {
