+++ homework_13_create-react-app/fish-shop/src/forms/showcase/Showcase.tsx
@@ -5,7 +5,7 @@ import {Products} from "../../components/products/Products";
 export class Showcase extends React.Component<any, any> {
   render() {
     return (
-      <main className="page-layout__main page-layout__main_theme_light">
+      <main className="page-layout__main">
         <Categories/>
         <Products/>
       </main>
